hxd: achieve an actual silent install/uninstall
How this is accomplished is a bit convoluted.  HxD uses the Inno Setup
installer system: https://jrsoftware.org/ishelp/index.php?topic=setupcmdline

For installation: /VERYSILENT is enough; /NORESTART is an extra nicety.
The rest is same as the existing model, just more verbose/clear PS-wise.

I'll add that, unlikey in the Chocolatey documentation, you DO NOT need
to specify a softwareName parameter -- this is a documentation mistake, as
the Chocolatey function itself ignores the parameter entirely.

For uninstallation: the situation is different.  Several things to note:

1. softwareName, used by Chocolatey's Get-UninstallRegistryKey function,
iterates over all the "DisplayName" entries in the uninstaller registry.
This is not well-documented and required review of the Chocolatey function
to determine behaviour.  In the case of HxD, we use a wildcard match,
because the string is "HxD Hex Editor 2.4" etc..  This parameter is very
important!

2. Get-UninstallRegistryKey returns an array of objects, since more than
one uninstall registry key may be found.  The $key.Count "stuff" is
to ensure that only one was found.

3. Inno Setup /VERYSILENT is not always enough; we need /SUPPRESSMSGBOXES
as well.  /NORESTART is explained above.

4. We use .skipAutoUninstaller here, since otherwise Chocolatey's
auto-installer runs regardless of you doing it yourself.  You might ask:
why not just let the auto-uninstaller do it?  The answer: because Inno
Setup's installer (uninstaller), when run, uses /SILENT unconditionally,
and doesn't let us specify other flags (it will error out if you try to do
things like /SILENT /VERYSILENT).  For the default string used, refer to
QuietUninstallString in the uninstaller registry.

Ticket https://github.com/Jackenmen/choco-auto/pull/7 documents this
story in full.