Several improvements, remove portable support
- Switch to using AU (chocolatey-au) for auto-updating
- Pre-package the plugin (.plgx file) itself as part of package building
- Add legal/ files
- Simplify and clean up install/uninstall scripts, plus fix a few bugs

Note: this package no longer supports portable KeePass installations.
If someone yells loud enough, I might spend the time to re-add the code
for it.  But because Get-BinRoot is gone in Chocolatey 2.x, we'd have to
defer to Get-ToolsLocation, which I do not believe is where keepass.portable
installs itself.  Would need help figuring it out.  :)